+++ provider/pkg/gen/utilities.go
@@ -20,7 +20,7 @@ func apiParamNameToSdkName(name string) string {
 	if strings.HasSuffix(name, "Id") {
 		name = inflector.Singularize(name[:len(name)-2])
 		switch name {
-		case "project", "location", "managedZone":
+		case "project", "location", "managedZone", "zone":
 			return name
 		case "datum":
 			return "dataId"
@@ -35,19 +35,62 @@ func apiParamNameToSdkName(name string) string {
 // In particular, it converts pluralized ID names to singular ID names.
 func apiPropNameToSdkName(name string) string {
 	switch name {
-	case "projectId", "locationId", "managedZoneId":
+	case "projectId", "locationId", "managedZoneId", "zoneId":
 		return name[:len(name)-2]
 	default:
 		return ToLowerCamel(name)
 	}
 }
 
-// nameRegexes is a list of regular expressions to match property formats in property descriptions.
-var nameRegexes = []*regexp.Regexp {
+// propertyFormatRegexes is a list of regular expressions to match property formats in property descriptions.
+var propertyFormatRegexes = []*regexp.Regexp{
 	// Example: "The resource name in the format `foo/*`"
 	regexp.MustCompile(`(?:Format|format|form|forms|formatted|pattern|Example|example|e\.g\.|such|Structured)(?: (?:of|is|as|like|will be))?[^\w]+(\w+(?:/[\w\{\}\*-\[\]]*)+)`),
 	// Example: "Must be in `projects/{project}/locations/{location}/triggers/{trigger}` format"
 	regexp.MustCompile(`in[^\w]+(\w+(?:/[\w\{\}\*-\[\]]*)+)[^\w]+format`),
+}
+
+// propertyPattern extracts the pattern to build property values using given property name (converted to SDK convention)
+// and description. Returns an empty string if no suitable pattern is found.
+// Currently, we support only patterns that end with the same property name, so not patterns that are a composition of
+// other properties, or more complex arrangements.
+func propertyPattern(sdkName, desc string, params codegen.StringSet) string {
+	if !strings.Contains(desc, "/") {
+		return ""
+	}
+
+	var match string
+	for _, nameRegex := range propertyFormatRegexes {
+		matches := nameRegex.FindStringSubmatch(desc)
+		if len(matches) == 2 {
+			match = matches[1]
+			break
+		}
+	}
+
+	parts := strings.Split(match, "/")
+	if len(parts) == 0 || len(parts)%2 == 1 {
+		return ""
+	}
+
+	var res []string
+	for i := 0; i < len(parts)-2; i += 2 {
+		part := parts[i]
+		paramName := apiParamNameToSdkName(part + "Id")
+		if !params.Has(paramName) {
+			return ""
+		}
+		res = append(res, fmt.Sprintf("%s/{%s}", part, paramName))
+	}
+
+	// The last segment has to match the property name.
+	lastProp := inflector.Singularize(parts[len(parts)-2])
+	if sdkName != lastProp {
+		return ""
+	}
+	res = append(res, fmt.Sprintf("%s/{%s}", parts[len(parts)-2], sdkName))
+
+	return strings.Join(res, "/")
 }
 
 // namePropertyPattern extracts the pattern to build the name property for the given resource's property specs.
@@ -64,7 +107,7 @@ func namePropertyPattern(inputProperties map[string]schema.PropertySpec) (string
 	}
 
 	var match string
-	for _, nameRegex := range nameRegexes {
+	for _, nameRegex := range propertyFormatRegexes {
 		matches := nameRegex.FindStringSubmatch(nameProp.Description)
 		if len(matches) == 2 {
 			match = matches[1]
@@ -82,7 +125,7 @@ func namePropertyPattern(inputProperties map[string]schema.PropertySpec) (string
 	}
 
 	var res []string
-	for i := 0; i < len(parts)-2; i+=2 {
+	for i := 0; i < len(parts)-2; i += 2 {
 		part := parts[i]
 		val := apiParamNameToSdkName(part + "Id")
 		if _, ok := inputProperties[val]; ok {